# Simplify coalesce operator