--- a/execution/exchange/coalesce.go
+++ b/execution/exchange/coalesce.go
@@ -80,28 +80,21 @@ func (c *coalesceOperator) Next(ctx context.Context) ([]model.StepVector, error)
 			if in == nil {
 				return
 			}
-			mu.RLock()
+			mu.Lock()
+			defer mu.Unlock()
+
 			if len(in) > 0 && out == nil {
-				mu.RUnlock()
-				mu.Lock()
-				if len(in) > 0 && out == nil {
-					out = c.pool.GetVectorBatch()
-					for i := 0; i < len(in); i++ {
-						out = append(out, c.pool.GetStepVector(in[i].T))
-					}
+				out = c.pool.GetVectorBatch()
+				for i := 0; i < len(in); i++ {
+					out = append(out, c.pool.GetStepVector(in[i].T))
 				}
-				mu.Unlock()
-			} else {
-				mu.RUnlock()
 			}
 
-			mu.Lock()
 			for i := 0; i < len(in); i++ {
 				out[i].Samples = append(out[i].Samples, in[i].Samples...)
 				out[i].SampleIDs = append(out[i].SampleIDs, in[i].SampleIDs...)
 				o.GetPool().PutStepVector(in[i])
 			}
-			mu.Unlock()
 			o.GetPool().PutVectors(in)
 		}(o)
 	}