@fpetkovski 2022-11-04
Good point, updated. I also converted the mutex and waitgroup to struct variables to reduce allocs.